Introduce WebViewV2 for WebPlatformPlayer
