+++ apps/ledger-live-desktop/package.json
@@ -61,6 +61,8 @@
     "@ledgerhq/types-cryptoassets": "workspace:^",
     "@ledgerhq/types-devices": "workspace:^",
     "@ledgerhq/types-live": "workspace:^",
+    "@ledgerhq/wallet-api-core": "^0.5.3",
+    "@ledgerhq/wallet-api-server": "^0.5.4",
     "@polkadot/react-identicon": "0.89.2",
     "@sentry/electron": "^4.0.0",
     "@sentry/node": "7.8.1",
